updated 2

+++ 0x0F-function_pointers/2-int_index.c
@@ -9,7 +9,7 @@
  *        (-1) - if no matches and size <= 0
  */
 
-int int_index(int *array, size_t size, int (*cmp)(int))
+int int_index(int *array, int size, int (*cmp)(int))
 {
 	unsigned int i;
